Reorganize file structure and some cleanup.
