+++ windfarm-extension/src/components/decorators/SensorDecorator.tsx
@@ -1,6 +1,6 @@
 import { DecorateContext, Decorator } from "@bentley/imodeljs-frontend";
-import { PowerMarker } from "./PowerMarker";
-import { SensorMarker } from "./SensorMarker";
+import { PowerMarker } from "../markers/PowerMarker";
+import { SensorMarker } from "../markers/SensorMarker";
 
 export class SensorDecorator implements Decorator {
   protected marker: SensorMarker;
