+++ windfarm-extension/src/components/decorators/WindDecorator.tsx
@@ -1,6 +1,6 @@
 import { DecorateContext, Decorator } from "@bentley/imodeljs-frontend";
-import { PowerMarker } from "./PowerMarker";
-import { WindMarker } from "./WindMarker";
+import { PowerMarker } from "../markers/PowerMarker";
+import { WindMarker } from "../markers/WindMarker";
 
 export class WindDecorator implements Decorator {
   protected marker: WindMarker;
